requirements.txt
python-dotenv
pandas
numpy
seaborn
certifi
pymongo[srv]==3.12
scikit-learn
scipy
pylance
pyYAML
-e .